Add function for removing counter event objects
This commit adds a utility function to the timestamped counter that can
trim the set of event objects stored for the counter. This is useful if
the number of event objects have gotten very large, but it is not
generally a good idea to use the function very often, because it naively
trims values without considering the impact to reporting.